Put accidental changes back the way they were.
