--- src/Highway.Data.Test/InMemory/BugTests/Customer.cs
@@ -1,17 +1,0 @@
-using System;
-
-namespace Highway.Data.Tests.InMemory.BugTests
-{
-    public class Customer
-    {
-        public void SetBusinessEntity(BusinessEntity businessEntity)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void SetEntityType(EntityType entityType)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
