--- src/Highway.Data.Test/InMemory/BugTests/EntityType.cs
@@ -1,9 +1,0 @@
-namespace Highway.Data.Tests.InMemory.BugTests
-{
-    public class EntityType
-    {
-        public string EntityTypeName { get; set; }
-
-        public int Id { get; set; }
-    }
-}
