--- src/Highway.Data.Test/InMemory/BugTests/FindFooById.cs
@@ -1,14 +1,0 @@
-using System.Linq;
-
-using Highway.Data.Tests.TestDomain;
-
-namespace Highway.Data.Tests.InMemory.BugTests
-{
-    public class FindFooById : Scalar<object>
-    {
-        public FindFooById(int i)
-        {
-            ContextQuery = c => c.AsQueryable<Foo>().SingleOrDefault(x => x.Id == i);
-        }
-    }
-}
