--- src/Highway.Data.Test/InMemory/BugTests/StatusType.cs
@@ -1,6 +1,0 @@
-namespace Highway.Data.Tests.InMemory.BugTests
-{
-    public class StatusType
-    {
-    }
-}
