+++ src/Highway.Data.Test/InMemory/BugTests/TestMultipleCommitsOrphansItems.cs
@@ -1,4 +1,6 @@
-﻿using FluentAssertions;
+﻿using System.Linq;
+
+using FluentAssertions;
 
 using Highway.Data.Contexts;
 using Highway.Data.Tests.TestDomain;
@@ -21,6 +23,14 @@ namespace Highway.Data.Tests.InMemory.BugTests
             var foo = repository.Find(new FindFooById(1));
 
             foo.Should().NotBeNull();
+        }
+    }
+
+    public class FindFooById : Scalar<object>
+    {
+        public FindFooById(int i)
+        {
+            ContextQuery = c => c.AsQueryable<Foo>().SingleOrDefault(x => x.Id == i);
         }
     }
 }
